Regenerated PlayerInput with correct class name. Improved PlayerController to use rigidbody and physics for movement.,
Bug/WIP: The characters Y-movement gets out of control when moving towards objects which the character can walk over/walk on.

Also TODO: have to fix the characters rotation based on the movement and have to fix/try how the characters movement feels if the character moves "north" instead of "north west" when pressing W etc.