Tried to implement a calculation for correct movement on slopes (not working at the moment)
The player can't move at all on slopes (either the force isn't enough or there is some other problem).

The slopes also allow for the player to start flying around, have to work a lot on the movement again.

Implemented a WIP movement speed limiter, which at the moment is a bit screwy as well (falling speed is affected and falling down feels slow).