WIP: Dash on the ground with corrected direction. Description for more info.
Dashing on a slope while facing downwards now doesn't push the player in the air.

Need to create a calculation for if the player goes off the slope that the remaining velocity's direction is updated to the correct direction instead of allowing the player to fly up in the air.