Removed unnecessary returns to Idle state. Created state handling for AbilityBlindingSand and AbilityInteract. More info ->
There were separate returns to Idle state from Moving and Dashing, which are now replaced with the baseline PlayerIdle as the problems which required the separate returns are solved.

AbilityBlindingSand and AbilityInteract states now work and require testing. (Also requires the box interacting to be fixed to test it further.)

Fixed spamming of playerstate changes and fixed groundcheck forcing the player to idle on almost every action and movement.

Added a test movable box to Level_2 (current spawn location) to test handling of Interact state.

Added correct UnityActions to PlayerStateManager and DebugManager and made them work with corresponding methods. (Ie. playerstate updates correctly and so does the playerstate)