Imlemented Damager.cs, Made changes to IHealth and Health and removed unnecessary and unused variables from enemies. Info ->
Also made changes to enemy projectile and it's despawning.
Made use of existing AliveTimer and made check ups so that the coroutine doesn't stay running if the projectile is disabled / destroyed.

Added functionality to PlayerSword to check if the player is attacking and if the player is attacking, then allow for the TriggerEnter to continue.