The EnemyProjectiles are now pooled and RangedEnemy can use them properly. More info ->
The RangedEnemy now has a cooldown routine which it uses to delay time between shots.

EnemyProjectile(s) speed can be set straight from the RangedEnemy, and the projectiles despawn on collision and/or after they have existed for their alive time.

Started work on enemy deaths in Health and EnemyHealth.

Created layer for EnemyProjectile and made changes to collision matrix.