Some folder hierarchy changes to UI prefab folder. Implemented HealthDisplay (needs testing). Changed Health.cs action from Action<int> to just an Action. More info ->
Noticed a bug with health pickups, the health pick ups try to increase the health of enemies as well.

Health display isn't probably the best or cleanest piece of code, but it works at least according to my own testing.

Added a placeholder image called HP_Empty and renamed the old HP image to HP_Full.