Reverted the change in Health.cs and instead made another way for the UIHealthManager to be able to listen to the players health change event. More info ->
First off I made the DecreaseHealth and IncreaseHealth in Health.cs public virtual ones, so that they can be overrided in PlayerHealth.cs.

Then in PlayerHealth.cs I created a new public static event Action which is called in the overridden Decrease- and IncreaseHealth, to update the UI health display.

Now the bug with every object with a Health script calling their OnHealthChanged when any other objects health changes is fixed.