The player can now pick up the rock with better animations. More info ->
Implemented a LiftableRockDropPoint for the player, so that the player can drop a LiftableRock a bit further away from the player, thus the player can't be pushed back by the rock.

Corrected a check up in GroundCheck to not allow going from InteractingIdle OR InteractingMove to InAir/Moving/Idle.

Applied changes to Player.prefab (addition of LiftableRockDropPoint)

LiftableRockDropPoint.cs only gets the transform.position of the LiftableRockDropPoint as a public property.